move video tests
